refactoring: fixed tests code style
